Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -7,17 +7,17 @@ int main() {
     scanf(" %c", &operator);
     switch (operator) {
         case '+':
-            printf("%.2f\n", (float)num1 + num2);
+            printf("%d\n", (int)num1 + num2);
             break;
         case '-':
-            printf("%.2f\n", (float)num1 - num2);
+            printf("%d\n", (int)num1 - num2);
             break;
         case '*':
-            printf("%.2f\n", (float)num1 * num2);
+            printf("%d\n", (int)num1 * num2);
             break;
         case '/':
             if (num2 != 0) {
-                printf("%.2f\n", (float)num1 / num2);
+                printf("%d\n", (int)num1 / num2);
             } else {
                 printf("error\n");
             }
